requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Hugging Face & SAM3
huggingface-hub>=0.20.0
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
ffmpeg-python>=0.2.0

# Computer Vision & Tracking
supervision>=0.18.0
# ultralytics>=8.1.0  # YOLO kullanılmıyor, SAM3 kullanılıyor
numpy>=1.24.0
Pillow>=10.0.0

# Object Tracking (Multi-player tracking)
filterpy>=1.4.5
scikit-learn>=1.3.0

# OCR for Jersey Number Recognition
paddlepaddle>=2.5.0
paddleocr>=2.7.0
# Alternative: easyocr>=1.7.0  # Uncomment if preferred over PaddleOCR

# Data Analysis
pandas>=2.1.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1

# YouTube Video Download
yt-dlp>=2023.12.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Social Media APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
requests>=2.31.0

# API (Optional - for future web service)
# fastapi>=0.104.0
# uvicorn>=0.24.0

